relay_mcp/requirements.txt
# RelayMCP Dependencies

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Flask dashboard
flask==3.0.0
requests==2.31.0

# Data & persistence
pyyaml==6.0.1

# Async support
asyncio==3.4.3

# Optional: OpenAI API (for GPT connector)
# openai==1.3.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Logging
python-json-logger==2.0.7
